Fix for color enhancement - cast to int

+++ classification/color_enhacement.py
@@ -9,7 +9,7 @@ class ColorEnhancement:
 
     def process_frame(self):
         (channel_b, channel_g, channel_r) = cv.split(self.img)
-        s = np.array(channel_b + channel_g + channel_r) + np.finfo(float).eps
-        v = np.dstack((channel_b-channel_g, channel_b-channel_r)).min(axis=2)
+        s = np.array(channel_b.astype(int) + channel_g.astype(int) + channel_r.astype(int)) + np.finfo(float).eps
+        v = np.dstack((channel_b.astype(int)-channel_g.astype(int), channel_b.astype(int)-channel_r.astype(int))).min(axis=2)
         output = np.dstack((v/s, np.zeros(v.shape))).max(axis=2)
         return output
